fix: - small fixes and changes - using conrrect useQuery hook for better typing assertions - removing uncessary component layer for product list - using formData to get form info instead of dedicated input ref
