+++ .eslintrc.json
@@ -34,7 +34,8 @@
     "semi": ["error", "always"],
     "linebreak-style": ["error", "unix"],
     "react/react-in-jsx-scope": "off",
-    "@typescript-eslint/no-unused-vars": ["error"]
+    "@typescript-eslint/no-unused-vars": ["error"],
+    "@typescript-eslint/no-explicit-any": "warn"
   },
   "settings": {
     "react": {
