+++ src/Hooks/useState.tsx
@@ -2,7 +2,6 @@ import { useState as useReactState } from "react";
 
 type NewState<T> = Partial<T> | ((newState: T) => void);
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 const useState = <T extends { [key: string]: any }>(
   defaultState: T,
 ): [T, (newState: NewState<T>) => void] => {
